Split up topics reducer into different files.
